FIX: lineHeight of ListItem icon
