+++ src/List/index.js.flow
@@ -32,8 +32,5 @@ type ListContextProps = {|
 
 export type ListContextType = React.Context<ListContextProps>;
 
-export type GetLineHeightToken = ({ theme: Theme, size: Size }) => string;
-
 declare export var ListContext: ListContextType;
 
-declare export var getLineHeightToken: GetLineHeightToken;
